add default float

+++ simulation.py
@@ -148,7 +148,7 @@ class Grid:
         state_array = stats.loc[:, ['supply', 'demand', 'arrival', 'expense', 'remaining_budget', 'unsatisfied_ratio']].values.reshape(1, 1, 10, 10, stats.shape[1] - 2).astype(np.float32)
         state_array = state_array.transpose(0, 1, 3, 2, 4).reshape(1, 1, 100, 6)
         self.state_memory.append(state_array)
-        return np.concatenate(self.state_memory, axis=1)
+        return np.concatenate(self.state_memory, axis=1).astype(np.float32)
 
     def get_last_satisfied_requests(self):
         return self.stats['satisfied_requests'].sum()
